requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pyyaml

numpy
scikit-learn

faiss-cpu
sentence-transformers
